update cython bindings for int8

+++ tocsin/reader.pxd
@@ -8,6 +8,6 @@ cdef extern from "<stdint.h>" nogil:
 
 cdef extern from "../fastp-bind/src/api.h":
     vector[string] fastq_reader(string fastq_filename)
-    vector[vector[int]] fastq_encoder(string fastq_filename, int length)
+    vector[vector[int8_t]] fastq_encoder(string fastq_filename, int length)
     vector[vector[int8_t]] fastq_encoder_alt(string fastq_filename,
                                               int length)
